pyproject.toml
[tool.poetry]
name = "photo_nametag"
version = "0.1.0"
description = ""
authors = ["YAMASHITA Yuta <proyuuta0618188188@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
face-recognition = "^1.3.0"
click = "^8.0.1"
pyexiv2 = "^2.6.3"
Pillow = "^8.3.1"
scikit-learn = "^0.24.2"
tqdm = "^4.61.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
photo_nametag = "photo_nametag.main:cli"
